+++ CHANGELOG.md
@@ -11,6 +11,7 @@ HumHub Changelog
 - Fix #7513: Fix priority of `dynamic.php`
 - Enh #7514: Update Redis example in `env.example`
 - Fix #7532: Prevent multiple execution of jobs canceled due to max. execution time.
+- Enh #7549: Allow logout from Mobile App
 
 1.17.2 (April 7, 2025)
 ----------------------